[squash] refactor classes
