+++ app/controllers/simplified-account/settings/worldpay-details/credentials/worldpay-credentials.controller.js
@@ -4,7 +4,7 @@ const paths = require('@root/paths')
 const { body, validationResult } = require('express-validator')
 const formatValidationErrors = require('@utils/simplified-account/format/format-validation-errors')
 const worldpayDetailsService = require('@services/worldpay-details.service')
-const { WorldpayCredential } = require('@models/gateway-account-credential/GatewayAccountCredential.class')
+const WorldpayCredential = require('@models/gateway-account-credential/WorldpayCredential.class')
 
 function get (req, res) {
   return response(req, res, 'simplified-account/settings/worldpay-details/credentials', {
